fix(zomboid): close client review ledger — fail-closed parser paths, vendored spec closure, narrowed lint exemption
Closes review ledger C-conf-3 (dictionary ITEMS section must be consumed
whole; unclosed or malformed records → corrupt_file), C-conf-4/C-qual-imp-1
(encoder errors surface as parse_error + exit 1; every emitted float is
checked finite), C-conf-5/C-qual-3 (a player without a descriptor →
corrupt_file naming the row instead of a nil dereference), C-conf-6/C-qual-4
(activeMods sorted; absent mods.txt → []), C-conf-8 (the pzdataspec import
closure is now vendored verbatim — 84 .ksy files — with a test that walks the
import graph), C-sec-imp-2/C-qual-1 (lint exemption narrowed to the sqlite and
blob decoders; gosec/govet no longer excluded), C-qual-2 (cell offsets must
start after the pointer array and cell-content start; interior rowid varints
bounded), C-qual-imp-2/3/4 (errors.Is, exact worn-item assertion, mixed
world-version fixture).

+++ plugins/zomboid/parser/blob/decode.go
@@ -294,7 +294,7 @@ func decodeContainer(r *reader) []InventoryItem {
 func decodeItemGroup(r *reader) InventoryItem {
 	var item InventoryItem
 	identical := r.count()
-	item.Count = int32(identical)
+	item.Count = identical
 	size := r.byteLen() // sized_blob.len_data
 	end := r.pos + size
 	item.RegistryID = r.u16()
